refactor: promptplayground to ee
