+++ web/src/ee/features/playground/page/components/ChatMessages.tsx
@@ -1,8 +1,8 @@
 import { PlusCircleIcon } from "lucide-react";
 
 import { Button } from "@/src/components/ui/button";
-import { ChatMessageComponent } from "@/src/features/playground/client/components/ChatMessageComponent";
-import type { MessagesContext } from "@/src/features/playground/client/components/Messages";
+import { ChatMessageComponent } from "@/src/ee/features/playground/page/components/ChatMessageComponent";
+import type { MessagesContext } from "@/src/ee/features/playground/page/components/Messages";
 import { ChatMessageRole } from "@langfuse/shared";
 import { useRef, useEffect } from "react";
