+++ web/src/ee/features/playground/page/components/GenerationOutput.tsx
