+++ web/src/ee/features/playground/page/components/PromptVariableComponent.tsx
