+++ web/src/ee/features/playground/page/utils/createEmptyMessage.ts
