+++ web/src/ee/features/playground/server/chatCompletionHandler.ts
@@ -11,8 +11,16 @@ import {
 } from "./validateChatCompletionBody";
 import { getCookieName } from "@/src/server/utils/cookies";
 import { env } from "@/src/env.mjs";
+import { getIsCloudEnvironment } from "@/src/ee/utils/getIsCloudEnvironment";
 
 export default async function chatCompletionHandler(req: NextRequest) {
+  if (!getIsCloudEnvironment()) {
+    return NextResponse.json(
+      { message: "This endpoint is available in Langfuse cloud only." },
+      { status: 501 },
+    );
+  }
+
   const token = await getToken({
     req,
     cookieName: getCookieName("next-auth.session-token"),
