+++ web/src/ee/features/playground/server/validateChatCompletionBody.ts
