+++ web/src/ee/utils/getIsCloudEnvironment.ts
@@ -1,0 +1,4 @@
+import { env } from "@/src/env.mjs";
+
+export const getIsCloudEnvironment = () => 
+ Boolean(env.NEXT_PUBLIC_LANGFUSE_CLOUD_REGION)
