--- web/src/features/playground/Readme.md
@@ -1,4 +1,0 @@
-NOTE: We may transition this feature from our MIT licensed repository to the
-a commercial License (ee folder) once we release a first stable version.
-Please consider this when planning long-term use and integration of this functionality into your projects.
-For more information see https://langfuse.com/docs/open-source
