+++ web/src/pages/project/[projectId]/evals/templates/new.tsx
@@ -1,6 +1,6 @@
 import Header from "@/src/components/layouts/header";
 import { EvalTemplateForm } from "@/src/features/evals/components/template-form";
-import { PlaygroundProvider } from "@/src/features/playground/client/context";
+import { PlaygroundProvider } from "@/src/ee/features/playground/page/context";
 import { evalLLMModels } from "@langfuse/shared";
 
 import { useRouter } from "next/router";
